requirements.txt
google-generativeai
fastapi
uvicorn
pydantic
dotenv
httpx
